Ported changes from ZDoom revision 417-426: - Fixed: Inventory.Icon printed error message for all the missing icons in Strife's teaser. - Fixed a few uninitialized variables warnings for GCC. - Turned on warning level 4 just to see what it would produce: a lot of warnings. At first, I was going to try and clean them all up. Then I decided that was a worthless cause and went about just acting on the ones that might actually be helpful:    C4189 (local variable is initialized but not referenced)    C4702 (unreachable code)    C4512 (assignment operator could not be generated) - Fixed: Commander Keen's death sequence had one state duplicated. - Fixed: Due to the changes for custom states the internal weapons shouldn't define a holdatk state unless it differs from the attack state. - Changed: MF2_BLASTED doesn't work well for players so I had to take it out of A_BlastRadius. - Fixed: When MF2_BLASTED was cleared MF2_SLIDE got cleared as well, no matter whether it was set by default or not. Now the MF2_SLIDE check checks both flags and the BlastRadius code doesn't set MF2_SLIDE anymore. - Added new flag MF5_NOBLOODDECALS that prevents bleeding actors from spawning blood decals. - Replaced several calls to S_GetSoundPlayingInfo with S_IsActorPlayingSomething because S_GetSoundPlayingInfo cannot properly resolve player and random sounds. - Fixed: S_IsActorPlayingSomething has to resolve aliases and player sounds. - Moved a_puzzleitems.cpp from g_hexen to g_shared
SVN r741 (trunk)

+++ src/d_netinfo.cpp
@@ -460,9 +460,9 @@ void D_SendServerFlagChange (const FBaseCVar *cvar, int bitnum, bool set)
 	namelen = (int)strlen (cvar->GetName ());
 
 	Net_WriteByte (DEM_SINFCHANGEDXOR);
-	Net_WriteByte (namelen);
+	Net_WriteByte ((BYTE)namelen);
 	Net_WriteBytes ((BYTE *)cvar->GetName (), namelen);
-	Net_WriteByte (bitnum | (set << 5));
+	Net_WriteByte (BYTE(bitnum | (set << 5)));
 }
 
 void D_DoServerInfoChange (BYTE **stream, bool singlebit)
